requirements.txt
autopep8==2.0.4
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
Pillow==10.1.0
PyAudio @ file:///C:/Users/Sistemas/Documents/GitHub/jarvis-python/PyAudio-0.2.11-cp311-cp311-win_amd64.whl#sha256=5092b46bff5d6687db817bde4a71ffc50dd316cc2e490ee81908f0c95ef2644f
PyAutoGUI==0.9.54
pycodestyle==2.11.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
python-decouple==3.8
pyttsx3==2.90
pytweening==1.0.7
pywhatkit==5.4
pywin32==306
requests==2.31.0
soupsieve==2.5
SpeechRecognition==3.10.0
urllib3==2.1.0
Werkzeug==3.0.1
wikipedia==1.4.0
